added Path as separate object + added ability to set scores using structure string

+++ src/graphs/Core/Interfaces/IVertex.cs
@@ -5,7 +5,7 @@ namespace Core.Interfaces
     public interface IVertex
     {
         string Name { get; }
-        IVertex AddEdge(string endVertexName, short weight = 0);
+        IVertex AddEdge(string endVertexName, int weight = 0);
         IReadOnlyList<IEdge> GetEdges();
     }
 }
